pyproject.toml
[project]
name = "fyp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "albumentations~=2.0.8",
    "grpcio-tools>=1.76.0",
    "grpcio>=1.76.0",
    "ipykernel~=6.17.1",
    "matplotlib~=3.10.0",
    "mypy-protobuf>=3.6.0",
    "numpy~=2.0.2",
    "opencv-python~=4.12.0.88",
    "pandas~=2.2.2",
    "roboflow~=1.2.11",
    "scipy~=1.16.2",
    "torchaudio>=2.8.0",
    "torchvision~=0.23.0",
    "torch~=2.8.0",
    "ultralytics~=8.3.218",
]

[tool.basedpyright]
include = [ "**/*.py" ]
exclude = [ "**/node_modules", "**/__pycache__" ]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
reportMissingImports = "error"
reportMissingTypeStubs = "error"
reportPrivateImportUsage = "none"
reportPrivateLocalImportUsage = "none"
